+++ templates/index.php
@@ -1,12 +1,3 @@
-<?php
-function getTime() {
-    $midnight = date_create("tomorrow midnight");
-    $today = date_create("now");
-    $diff = date_diff($today, $midnight);
-    $currentDiff = date_interval_format($diff, "%h<span>:</span>%I");
-    return $currentDiff;
-}
-?>
 <section class="promo">
     <h2 class="promo__title">Нужен стафф для катки?</h2>
     <p class="promo__text">На нашем интернет-аукционе ты найдёшь самое эксклюзивное сноубордическое и горнолыжное снаряжение.</p>
